feat: Add Docker Build and Push Beta workflow
This commit adds a new workflow, 'Docker Build and Push Beta', which automates the build and push process for a Docker image. The workflow is triggered on a push to the 'master' branch and manually through workflow_dispatch.

The workflow consists of multiple steps including clearing digests, building the image for GHCR, exporting and uploading the digest as an artifact, and merging the digest for GHCR. The workflow utilizes Docker Buildx, Docker meta, and Docker login to perform these steps.

The new workflow replaces the deprecated 'continuous-integration-workflow.yml'.